requirements.txt
flask==2.0.1
werkzeug==2.0.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
requests==2.31.0
textblob==0.15.3
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
pydantic==1.10.13
httpx==0.24.1
pandas==2.0.3
lxml==4.9.3
flask-cors 